refactor: unify manager interfaces for easier mocks
